backend/requirements.txt
fastapi
uvicorn
python-multipart
# tensorflow <- Removed due to Python 3.14 incompatibility
torch
torchvision
pillow
numpy
pytest
httpx
